+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/IJdbcTable.java
@@ -48,7 +48,7 @@ import java.util.Iterator;
  * Extension of the {@link ITable} specially dedicated to the JDBC databases thanks to the extension of the
  * {@link ResultSet} interface. It also extends the {@link IWhereBuilderOrOptionBuilder} for the SQL requesting
  */
-public interface IJdbcTable extends ITable, ResultSet, IWhereBuilderOrOptionBuilder {
+public interface IJdbcTable extends ITable<Object>, ResultSet, IWhereBuilderOrOptionBuilder {
 
     /** {@link String} name of the metadata property */
     String META_PROPERTY = "meta";