+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/ISpatialTable.java
@@ -48,7 +48,7 @@ import java.util.Map;
  * @author Erwan Bocher (CNRS)
  * @author Sylvain PALOMINOS (UBS 2018-2019)
  */
-public interface ISpatialTable extends ITable {
+public interface ISpatialTable extends ITable<Object> {
 
     /**
      * Returns the {@link Geometry} value of the given column at the current row.