+++ data-manager/jdbc/src/main/java/org/orbisgis/orbisdata/datamanager/jdbc/JdbcTable.java
@@ -237,7 +237,7 @@ public abstract class JdbcTable extends DefaultResultSet implements IJdbcTable, 
     }
 
     @Override
-    public Collection<String> getColumnNames() {
+    public Collection<String> getColumns() {
         try {
             return JDBCUtilities
                     .getFieldNames(getResultSetLimit(0).getMetaData())
@@ -251,10 +251,10 @@ public abstract class JdbcTable extends DefaultResultSet implements IJdbcTable, 
     }
 
     @Override
-    public Map<String, String> getColumns(){
+    public Map<String, String> getColumnsTypes(){
         Map<String, String> map = new HashMap<>();
-        getColumnNames().forEach((name) -> {
-            map.put(name, getColumnsType(name));
+        getColumns().forEach((name) -> {
+            map.put(name, getColumnType(name));
         });
         return map;
     }
@@ -284,7 +284,7 @@ public abstract class JdbcTable extends DefaultResultSet implements IJdbcTable, 
     }
 
     @Override
-    public String getColumnsType(String columnName){
+    public String getColumnType(String columnName){
         if(!hasColumn(columnName)){
             return null;
         }
@@ -304,7 +304,7 @@ public abstract class JdbcTable extends DefaultResultSet implements IJdbcTable, 
 
     @Override
     public boolean hasColumn(String columnName){
-        return getColumnNames().contains(TableLocation.capsIdentifier(columnName, getDbType().equals(DataBaseType.H2GIS)));
+        return getColumns().contains(TableLocation.capsIdentifier(columnName, getDbType().equals(DataBaseType.H2GIS)));
     }
 
     @Override
@@ -382,24 +382,24 @@ public abstract class JdbcTable extends DefaultResultSet implements IJdbcTable, 
     }
 
     @Override
-    public ITable columns(String... columns){
+    public IJdbcTable columns(String... columns){
         WhereBuilder builder = new WhereBuilder(getQuery(columns), getJdbcDataSource());
         if(isSpatial()){
-            return builder.getSpatialTable();
+            return (IJdbcTable)builder.getSpatialTable();
         }
         else {
-            return builder.getTable();
+            return (IJdbcTable)builder.getTable();
         }
     }
 
     @Override
-    public ITable columns(List<String> columns){
+    public IJdbcTable columns(List<String> columns){
         WhereBuilder builder = new WhereBuilder(getQuery(columns.toArray(new String[0])), getJdbcDataSource());
         if(isSpatial()){
-            return builder.getSpatialTable();
+            return (IJdbcTable)builder.getSpatialTable();
         }
         else {
-            return builder.getTable();
+            return (IJdbcTable)builder.getTable();
         }
     }
 
@@ -485,7 +485,7 @@ public abstract class JdbcTable extends DefaultResultSet implements IJdbcTable, 
         if(propertyName.equals(META_PROPERTY)){
             return getMetaData();
         }
-        Collection<String> columns = getColumnNames();
+        Collection<String> columns = getColumns();
         if(columns!= null &&
                 (columns.contains(propertyName.toLowerCase()) || columns.contains(propertyName.toUpperCase()))
                 || "id".equalsIgnoreCase(propertyName)) {
@@ -525,7 +525,7 @@ public abstract class JdbcTable extends DefaultResultSet implements IJdbcTable, 
             else{
                 return this;
             }
-            Collection<String> columnNames = getColumnNames();
+            Collection<String> columnNames = getColumns();
 
             printer.startTable(ASCII_COLUMN_WIDTH, columnNames.size());
             printer.appendTableTitle(this.getName());