+++ data-manager/jdbc/src/test/groovy/org/orbisgis/orbisdata/datamanager/jdbc/GroovyPostGISTest.groovy
@@ -255,7 +255,7 @@ class GroovyPostGISTest {
                 CREATE TABLE postgis (id int, the_geom geometry(point, 0));
                 INSERT INTO postgis VALUES (1, 'POINT(10 10)'::GEOMETRY), (2, 'POINT(1 1)'::GEOMETRY);
         """)
-        assertEquals("id,the_geom", postGIS.getSpatialTable("postgis").columnNames.join(","))
+        assertEquals("id,the_geom", postGIS.getSpatialTable("postgis").columns.join(","))
     }
 
     @Test