+++ data-manager/jdbc/src/test/java/org/orbisgis/orbisdata/datamanager/jdbc/JdbcTableTest.java
@@ -321,7 +321,7 @@ public class JdbcTableTest {
     }
 
     /**
-     * Test the {@link JdbcTable#getColumnNames()} method.
+     * Test the {@link JdbcTable#getColumns()} method.
      */
     @Test
     public void testGetColumnNames(){
@@ -331,7 +331,7 @@ public class JdbcTableTest {
         colList.add(TableLocation.capsIdentifier(COL_ID, true));
         colList.add(TableLocation.capsIdentifier(COL_VALUE, true));
         colList.add(TableLocation.capsIdentifier(COL_MEANING, true));
-        assertEquals(colList, getTable().getColumnNames());
+        assertEquals(colList, getTable().getColumns());
     }
 
     /**
@@ -514,22 +514,22 @@ public class JdbcTableTest {
     }
 
     /**
-     * Test the {@link JdbcTable#getColumnsType(String)} method.
+     * Test the {@link JdbcTable#getColumnType(String)} method.
      */
     @Test
     public void testGetColumnsType() {
-        assertEquals("GEOMETRY", getTable().getColumnsType(COL_THE_GEOM));
-        assertEquals("INTEGER", getTable().getColumnsType(COL_ID));
-        assertEquals("VARCHAR", getTable().getColumnsType(COL_MEANING));
-        assertNull(getTable().getColumnsType("NOT_A_COLUMN"));
+        assertEquals("GEOMETRY", getTable().getColumnType(COL_THE_GEOM));
+        assertEquals("INTEGER", getTable().getColumnType(COL_ID));
+        assertEquals("VARCHAR", getTable().getColumnType(COL_MEANING));
+        assertNull(getTable().getColumnType("NOT_A_COLUMN"));
     }
 
     /**
-     * Test the {@link JdbcTable#getColumns()} method.
+     * Test the {@link JdbcTable#getColumnsTypes()} method.
      */
     @Test
     public void testGetColumns() {
-        Map<String, String> map = getTable().getColumns();
+        Map<String, String> map = getTable().getColumnsTypes();
         String[] keys = {COL_THE_GEOM2.toUpperCase(), COL_THE_GEOM, COL_ID, COL_VALUE, COL_MEANING};
         String[] values = {"POINTZ", "GEOMETRY", "INTEGER", "DOUBLE", "VARCHAR"};
         assertArrayEquals(keys, map.keySet().toArray());