+++ process-manager/process/src/test/groovy/org/orbisgis/orbisdata/processmanager/process/TestProcess.groovy
@@ -163,7 +163,7 @@ class TestProcess {
             title "With database"
             inputs inputA: ITable
             outputs outputA: String
-            run { inputA -> [outputA: inputA.columnNames] }
+            run { inputA -> [outputA: inputA.columns] }
         })